Support X-FORWARDED-FOR and X-FORWARDED-PROTO
On some platforms, like Heroku in this case, the way routing works is
that all requests appear to be from localhost *without* SSL. This is
just because TLS is terminated way before the app server.

Headers are set with the original request information, though. In this
case we care about X-FORWARDED-FOR, which is the original IP address of
the request sender, and X-FORWARDED-PROTO, which is the protocol used
(http/https).

If the headers don't exist (like they wouldn't locally) it falls back to
the normal Express `req.connection` data.