Allow sorting by `createdAt` ascending (#508)
When attempting to sort by the `createdAt` column, we would still apply
the default sort ordering of `-createdAt`, meaning that we would specify
a sort order of `createdAt,-createdAt`. When Parse-Server gets a sort
order in which the same field appears more than once, it will use the
last entry for the field in its sorting. This means that the sort the
user has selected is overwritten by the default sort we apply.

We now check to see if the sort we are applying is already on the
`createdAt` field, and if so we do not apply the default ordering.

Fixes #447.