Added support for reverse proxy
Currently the app.js assumes that all "non-127.0.0.1" requests are insecure if they are not HTTPS. 

However, this is an incorrect assumption as the app could be behind a reverse proxy. This is specially important when you use Docker to deploy an nginx that manages HTTPS and routes the request to Parse. In that case, the remote IP is in private range of 172.17.0.1 or 192.168.0.1.